Изменение long на int

+++ src/main/java/gs/konick/model/Role.java
@@ -4,22 +4,22 @@ public enum Role {
     CLIENT(1, "client"),
     MANAGER(2, "manager");
 
-    private long id;
+    private int id;
     private String role;
 
-    public static Role getRoleById(long id) {
+    public static Role getRoleById(int id) {
         for (Role r: values()) {
             if (r.id == id) return r;
         }
         throw new IllegalArgumentException("Вашего id " + id + " нет в доступных для role");
     }
 
-    Role(long id, String role) {
+    Role(int id, String role) {
         this.id = id;
         this.role = role;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
